renamed component RandomRecipe to Recipe
